se extraen del mw las funcionalidades de reloj vectorial y corte consistente, para utilizarlos como bibliotecas separadas se optimizaron las operaciones en modo automatico, unificando aquello que se repetía en todos los servidores

+++ scripts/levantar_bd.js
@@ -1,4 +1,4 @@
-var shell_ejec = require('./shell')
+var shell_ejec = require('./shell_con_cb')
 var async = require('async');
 var MongoClient = require('mongodb').MongoClient;
 var mongo_url = require("../cfg.json").mongo.url;
@@ -18,33 +18,40 @@ if (instancia_bd.length == 0 ||
 
 var operaciones = [
   function(callback) {
-    console.log("levantando bd");
+    console.log("Levantando DBMS mongo");
 
     // 3.0 pc trabajo o 3.4 pc casa y note
     var version_mongo = "3.4";
     // servidor base de datos (hay que crear carpeta C:\data\db)
-    shell_ejec("start ventana /K \"c:\ && cd Program Files && cd MongoDB && cd Server && cd " + version_mongo +  " && cd bin && mongod.exe\"");
 
-    setTimeout(function() { callback(null); }, 8000);
+    shell_ejec.execCommand("start ventana /K \"c:\ && cd Program Files && cd MongoDB && cd Server && cd " + version_mongo +  " && cd bin && mongod.exe\"", function (returnvalue) {
+      callback(returnvalue);
+    });
   },
   function(callback) {
 
-    if(instancia_bd == "nueva"){
-
-      console.log("limpiando base de datos");
+    console.log("Conectando a base de datos");
+    var id = setInterval(function(){
       MongoClient.connect(mongo_url, function(err, db) {
-        db.dropDatabase();
-        setTimeout(function() { process.exit(1); }, 1000);
-      });
-    }
-    if(instancia_bd == "actual"){
-      console.log("usando base de datos actual");
-    }
 
+        if(db !== null){
+          clearInterval(id);
+
+          if(instancia_bd == "nueva"){
+            console.log("Limpiando colecciones en base de datos");
+            db.dropDatabase();
+          }
+
+          if(instancia_bd == "actual"){
+            console.log("Se continua desde ultimo estado de base de datos");
+          }
+          process.exit(1);
+        }
+      });
+    }, 500);
     callback(null);
   }
 ];
 
 async.waterfall(operaciones, function (err, evento) {
-  console.log("todos los servidores activos");
 });
